Increases function name hitbox size
Previously function names were displayed inline rather than block, which made their hitbox only as wide as the word itself. This made it really hard to click on functions like 'is' and 'I'. Now it is displayed as a block element so they all take as much width as possible.